+++ packages/compiler/src/index.ts
@@ -1,5 +1,6 @@
-export { analyzeServerImports, isServerImportPath } from "./analyzer";
+export { analyzeServerImports, extractExportedFunctions, isServerImportPath } from "./analyzer";
 export type { ServerImport } from "./analyzer";
+export { generateHMRBlock } from "./hmr";
 export { shadejs } from "./plugin";
 export { generateRPCStub, getRPCRoutePath } from "./rpc-gen";
 export { generateProductionServer } from "./server-build";